pyproject.toml
[tool.poetry]
name = "verif-ai backend"
version = "0.1.0"
description = "Backend REST API for Verif.AI"
authors = ["ctagard <ctagard19@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "verif"}]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
django-environ = "^0.11.2"
drf-yasg = "^1.21.7"
requests = "^2.31.0"
django = "^5.0.2"
celery = "^5.3.6"
psycopg2-binary = "^2.9.9"
redis = "^5.0.3"
gunicorn = "^21.2.0"
django-cors-headers = "^4.3.1"
langchain = "^0.1.14"
langchain-community = "^0.0.31"
langchain-core = "^0.1.38"
langgraph = "^0.0.30"
pandas = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
